python_backend/requirements.txt
flask
torch
torchvision
Pillow
numpy
scikit-learn
scikit-image
opencv-python
